SEPP: fix db size calculation (#16364)
